+++ TREnvironmentEditor/Model/Types/Entities/EMMovePickupFunction.cs
@@ -169,7 +169,7 @@ public class EMMovePickupFunction : BaseEMFunction
         else
         {
             // Only look for the types we are interested in.
-            matchingEntities = entities.FindAll(e => Types.Contains(e.TypeID));
+            matchingEntities = entities.FindAll(e => Types.Contains((short)e.TypeID));
         }
 
         foreach (TR3Entity match in matchingEntities)
