+++ TRLevelControl/IO/TRLevelWriter.cs
@@ -159,7 +159,7 @@ public class TRLevelWriter : BinaryWriter
 
     public void Write(TR3Entity entity)
     {
-        Write(entity.TypeID);
+        Write((short)entity.TypeID);
         Write(entity.Room);
         Write(entity.X);
         Write(entity.Y);
