+++ TRRandomizerCore/Randomizers/TR3/TR3SecretRewardRandomizer.cs
@@ -43,12 +43,12 @@ public class TR3SecretRewardRandomizer : BaseTR3Randomizer
 
         foreach (int rewardIndex in secretMapping.RewardEntities)
         {
-            if (level.Data.Entities[rewardIndex].TypeID == (short)TR3Type.SaveCrystal_P)
+            if (level.Data.Entities[rewardIndex].TypeID == TR3Type.SaveCrystal_P)
             {
                 continue;
             }
 
-            level.Data.Entities[rewardIndex].TypeID = (short)stdItemTypes[_generator.Next(0, stdItemTypes.Count)];
+            level.Data.Entities[rewardIndex].TypeID = stdItemTypes[_generator.Next(0, stdItemTypes.Count)];
         }
     }
 }
